Forgot to add some files.
Also, changing the means that grades are tracked so that univ_oahpa.views can
be just a bit cleaner. Grades are tracked by decorating a view function, and
fancy names for the logs are specified in the views file.

Some changes to courses.views.trackGrade are probably required to make sure
that the log names for the game type (e.g., leksa book and language pair), but
I won't be able to do this until I'm done installing my test database.